# Optimize Pd::Workshop#potential_organizers query