--- a/dashboard/app/controllers/api/v1/pd/workshops_controller.rb
+++ b/dashboard/app/controllers/api/v1/pd/workshops_controller.rb
@@ -194,7 +194,9 @@ class Api::V1::Pd::WorkshopsController < ::ApplicationController
 
   # Users who could be re-assigned to be the organizer of this workshop
   def potential_organizers
-    render json: @workshop.potential_organizers
+    render json: @workshop.potential_organizers.pluck(:name, :id).map do |name, id|
+      {label: name, value: id}
+    end
   end
 
   private