--- a/dashboard/app/models/pd/workshop.rb
+++ b/dashboard/app/models/pd/workshop.rb
@@ -56,6 +56,8 @@ class Pd::Workshop < ActiveRecord::Base
   has_many :enrollments, class_name: 'Pd::Enrollment', dependent: :destroy, foreign_key: 'pd_workshop_id'
   belongs_to :regional_partner
 
+  has_many :regional_partner_program_managers, source: :program_managers, through: :regional_partner
+
   before_save :process_location, if: -> {location_address_changed?}
   auto_strip_attributes :location_name, :location_address
 
@@ -688,39 +690,30 @@ class Pd::Workshop < ActiveRecord::Base
   end
 
   # Users who could be re-assigned to be the organizer of this workshop
+  # @return [ActiveRecord::Relation]
   def potential_organizers
-    potential_organizer_ids = []
-    potential_organizers = []
-
-    # if there is a regional partner, only that partner's PMs can become the organizer
-    # otherwise, any PM can become the organizer
-    if regional_partner
-      regional_partner.program_managers.each do |pm|
-        potential_organizers << {label: pm.name, value: pm.id}
-      end
-    else
-      UserPermission.where(permission: UserPermission::PROGRAM_MANAGER).pluck(:user_id)&.map do |user_id|
-        potential_organizer_ids << user_id
-      end
-    end
-
-    # any CSF facilitator can become the organizer of a CSF workshhop
-    if course == Pd::Workshop::COURSE_CSF
-      Pd::CourseFacilitator.where(course: Pd::Workshop::COURSE_CSF).pluck(:facilitator_id)&.map do |user_id|
-        potential_organizer_ids << user_id
-      end
-    end
+    user_queries = []
 
     # workshop admins can become the organizer of any workshop
-    UserPermission.where(permission: UserPermission::WORKSHOP_ADMIN).pluck(:user_id)&.map do |user_id|
-      potential_organizer_ids << user_id
+    organizer_roles = [UserPermission::WORKSHOP_ADMIN]
+
+    if regional_partner_id
+      # if there is a regional partner, only that partner's PMs can become the organizer
+      user_queries << regional_partner_program_managers
+    else
+      # otherwise, any PM can become the organizer
+      organizer_roles << UserPermission::PROGRAM_MANAGER
     end
 
-    User.where(id: potential_organizer_ids).pluck(:name, :id)&.map do |name, id|
-      potential_organizers << {label: name, value: id}
+    user_queries << User.joins(:permissions).merge(UserPermission.where(permission: organizer_roles))
+
+    # any CSF facilitator can become the organizer of a CSF workshop
+    if course == Pd::Workshop::COURSE_CSF
+      user_queries << User.joins(:courses_as_facilitator).merge(Pd::CourseFacilitator.where(course: Pd::Workshop::COURSE_CSF))
     end
 
-    potential_organizers
+    # Combine multiple queries into single result set.
+    user_queries.inject(:union)
   end
 
   def can_user_delete?(user)