# Description

<!--
  A summary of the change, including any relevant motivation and context.

  If relevant, include a description both of the existing behavior and of the new behavior.
-->

Optimizes `Pd::Workshop#potential_organizers` logic using several `JOIN`s plus [`UNION`](https://dev.mysql.com/doc/refman/5.7/en/union.html) to reduce to a single ActiveRecord query.

Adds the [`active_record_union`](https://github.com/brianhempel/active_record_union) gem to support combining multiple `ActiveRecord::Relation` scopes via `UNION` into another `ActiveRecord::Relation` representing one SQL query.

In the process I changed the signature slightly, so that the model method returns a general-purpose `ActiveRecord::Relation<User>` representing the query, while the controller uses `#pluck` to pick specific values into a hash for the JSON response. I think this follows the model/controller separation a bit more cleanly.

### Background

This action is currently slow, in part due to an incorrect query plan selection that is possibly due to [MySQL Bug#81341](https://bugs.mysql.com/bug.php?id=81341). This optimization works around this bug by changing the type of query performed (using `JOIN` and `UNION` instead of `SELECT ... WHERE id IN (1, 2, 3, ...)`.

<!-- ### Privacy -->
<!-- ### Security -->
<!-- ### Caching -->
<!-- ### Testing -->
<!-- ### Deployment strategy -->
<!-- ### Future work -->

## Links

<!--
  Any relevant links to external resources; ie, specification documents, jira
  items, related PRs, honeybadger errors, etc
-->

- https://dev.mysql.com/doc/refman/5.7/en/union.html
- https://github.com/brianhempel/active_record_union

## Testing story

Updated the existing unit test covering this logic.

Query/performance comparison (running `pd_workshop.potential_organizers`):

Before:
```
(6.9ms)  SELECT `user_permissions`.`user_id` FROM `user_permissions` WHERE
 `user_permissions`.`permission` = 'program_manager'

(1.3ms)  SELECT `pd_course_facilitators`.`facilitator_id` FROM `pd_course_facilitators` WHERE
 `pd_course_facilitators`.`course` = 'CS Fundamentals'

(0.4ms)  SELECT `user_permissions`.`user_id` FROM `user_permissions` WHERE
 `user_permissions`.`permission` = 'workshop_admin'

(7495.4ms)  SELECT `users`.`name`, `users`.`id` FROM `users` WHERE `users`.`deleted_at` IS NULL
 AND `users`.`id` IN (#, #, #, #, #, ... etc ...)
```

Running last query directly: `503 rows in set (7.68 sec)`

After:
```
(28.4ms)  SELECT `users`.`name`, `users`.`id` FROM ( (SELECT `users`.* FROM `users` INNER JOIN
 `user_permissions` ON `user_permissions`.`user_id` = `users`.`id` WHERE `users`.`deleted_at`
 IS NULL AND `user_permissions`.`permission` IN ('workshop_admin', 'program_manager'))
 UNION (SELECT `users`.* FROM `users` INNER JOIN `pd_course_facilitators` ON
 `pd_course_facilitators`.`facilitator_id` = `users`.`id` WHERE `users`.`deleted_at` IS NULL AND
 `pd_course_facilitators`.`course` = 'CS Fundamentals') ) `users`
```

Running last query directly: `503 rows in set (0.03 sec)`